fix: minor dataset typo fixes

+++ apis/dataset_metadata/dataset_related_item.py
@@ -382,6 +382,11 @@ class RelatedItemTitlesDelete(Resource):
         if not is_granted("dataset_metadata", study_obj):
             return "Access denied, you can not make any change in dataset metadata", 403
         dataset_title_ = model.DatasetRelatedItemTitle.query.get(title_id)
+        if dataset_title_.type == "MainTitle":
+            return (
+                "Main Title type can not be deleted",
+                403,
+            )
         model.db.session.delete(dataset_title_)
         model.db.session.commit()
         return Response(status=204)
